+++ com.unity.shadergraph/Editor/Data/Nodes/Utility/SubGraphNode.cs
@@ -229,6 +229,9 @@ namespace UnityEditor.ShaderGraph
                     case PropertyType.Matrix4:
                         slotType = SlotValueType.Matrix4;
                         break;
+                    case PropertyType.SamplerState:
+                        slotType = SlotValueType.SamplerState;
+                        break;
                     default:
                         throw new ArgumentOutOfRangeException();
                 }
@@ -274,6 +277,14 @@ namespace UnityEditor.ShaderGraph
                     var tProp = prop as CubemapShaderProperty;
                     if (tSlot != null && tProp != null)
                         tSlot.cubemap = tProp.value.cubemap;
+                }
+                // copy default for gradient for niceness
+                else if (slotType == SlotValueType.Gradient && propType == PropertyType.Gradient)
+                {
+                    var tSlot = slot as GradientInputMaterialSlot;
+                    var tProp = prop as GradientShaderProperty;
+                    if (tSlot != null && tProp != null)
+                        tSlot.value = tProp.value;
                 }
                 AddSlot(slot);
                 validNames.Add(id);
@@ -356,7 +367,7 @@ namespace UnityEditor.ShaderGraph
             if (referencedGraph == null)
                 return;
 
-            referencedGraph.CollectShaderProperties(visitor, generationMode);
+            referencedGraph.CollectSubgraphProperties(visitor, generationMode);
         }
 
         public override void CollectPreviewMaterialProperties(List<PreviewProperty> properties)
@@ -404,7 +415,9 @@ namespace UnityEditor.ShaderGraph
                 // Generate arguments... first INPUTS
                 var arguments = new List<string>();
                 foreach (var prop in referencedGraph.properties)
+                {
                     arguments.Add(string.Format("{0}", prop.GetPropertyAsArgumentString()));
+                }
 
                 // now pass surface inputs
                 arguments.Add(string.Format("{0} IN", graphContext.graphInputStructName));
@@ -426,10 +439,10 @@ namespace UnityEditor.ShaderGraph
                     foreach (var node in nodes.OfType<AbstractMaterialNode>())
                     {
                         if (node is IGeneratesBodyCode)
-                            (node as IGeneratesBodyCode).GenerateNodeCode(bodyGenerator, graphContext, generationMode);
+                            (node as IGeneratesBodyCode).GenerateNodeCode(bodyGenerator, graphContext, GenerationMode.ForReals);
                     }
 
-                    outputNode.RemapOutputs(bodyGenerator, generationMode);
+                    outputNode.RemapOutputs(bodyGenerator, GenerationMode.ForReals);
 
                     s.Append(bodyGenerator.GetShaderString(1));
                 }